test: one test file for audioFeatures
audioFeaturesDrawer.test.ts was named for the work that introduced it
rather than the module it covers; nothing in it is drawer specific, and
the album page player runs the same code.
